Conditionally replace deprecated url import
The django.conf.urls.url() alias of django.urls.re_path() is
deprecated as of Django 3.1 and will be removed in 4.0 (which is
currently in alpha).

This commit changes the import to use the orignal alias for versions
before 3.1 and uses django.urls.re_path directly for newer versions.

See also:
https://docs.djangoproject.com/en/3.2/releases/3.1/#id2

Closes #41