Pass linked-field URL params to openModal
The ChooserWidgetFactory.openModal function (as introduced in #64 and used by https://github.com/wagtail/wagtail-multiple-chooser-panel/) did not work in conjunction with LinkedFieldMixin, because the logic for gathering linked field data into URL params was only in place for the widget's openModal method, not the factory's. Refactor it to be available on both.