Fix tests to account for search indexing happening as a background task
As of https://github.com/wagtail/wagtail/pull/12787 in Wagtail 6.4, search indexing ordinarily happens at the end of the transaction where the objects are created/updated. Since tests are wrapped in transactions, we need to explicitly invoke the oncommit callbacks via `self.captureOnCommitCallbacks(execute=True)` to ensure that the changes we make to the database are reflected in search results.